pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "realenv-ai"
version = "1.0.0"
description = "Elite Customer Support AI Agent Environment"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Vishnu" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi",
    "uvicorn",
    "groq",
    "openai>=2.7.2",
    "python-dotenv",
    "pyyaml",
    "pydantic",
    "requests",
    "streamlit",
    "pandas",
    "numpy",
    "jinja2",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "app:main"

[tool.setuptools.package-dir]
"" = "server"

[project.urls]
"Homepage" = "https://huggingface.co/spaces/TheVishnu/realenvai"
"Bug Tracker" = "https://github.com/Vishnu180806/realenv-ai/issues"
